Handle auth failures for user login
Previously, if a user authentication message came back from Simperium in
a failure mode, it received a string error message which failed to parse
as JSON in `user.js`. This resulted in an uncaught exception hidden
inside the carrying promise.

Now, those error messages are appropriately caught and passed up inside
a new `Error`, which then leads to the promise `reject()`'ing instead of
silently failing.

Notable Changes:
 - Pinned `eslint@2.2` in the Tavis CI config to prevent a bug which
   appeared with a recent release of `babel-eslint`
 - No longer supporting earlier than node 4

Props to @beaucollins for his work adding unit tests and making this
behave properly