Revisions: Increase count
The default revision collector grabs at most 30 revisions of an object
from a bucket but the default numbers for the revision count and
revision (ten) count are 60 and 100, respectively.

This patch bumps up the number of revisions collected so that for apps
without custom settings they can fetch all of the saved revisions.

More thorough work is needed because we have a tradeoff between revision
counts and latency in fetching them. If we could return the list of
revisions repeatedly as new ones come in this would not be a problem.