Refactor and add timeout
Since we're opening up more revisions I realize that this could result
in delays getting the revisions that could cause pain in the app. This
revision sets a new hard timeout at ten seconds so that if for some
reason we haven't been able to load all of the requested revisions we
can at least push out the ones we did get in that time.