Refactor again to use sequential fetch process
The original refactor was trying to do as little as possible to improve
the revision fetching but more work was needed to be more reliable. In
this round we're starting a new sequential process to try and quickly
get all avabilable revisions:

 - Launch an initial salvo with likely available versions
 - After this salvo comes back start sequentially requesting older
   revisions

We're trying to get back to version 1 here and this ends up spewing out
a bunch of requests for versions which don't exist. However, without
getting much deeper into way revisions are handled we can't easily
optimize this away. For now we'll just take the loss knowing that a
bunch of extra requests over the websocket will still be a small amount
of data relative to the note information.

This approach relies on a rolling timeout which keeps on being deferred
as long as valid revisions keep coming in. Once no more revisions come
back within the given delay we finish the process and consider it "good
enough of an attempt," returning the accumulated results to the
application.