See: Automattic/simplenote-electron/issues/1579
Discovered an issue where the local Simperium client was re-sending
changes that had already been accepted by the server. This was caused
by not recognizing the duplicate-change error and also by not
acknowledging that a queued change had come back from the server in
a change list.

In this patch we're adding the duplicate-change handler to remove
a change from the app's queue when the server tells us that it has
already been applied.

After this change we should see a reduction in duplicate-change load.