Refactor: Clarify names in `updateObjectVersion`
When reading through `updateObjectVersion` I was having trouble
understanding how the function should flow and what the variables held.
In this change I've replaced mutable variables with `const`s and renamed
many of the variables to focus more on their role than on their essence.

The function flow has also been tossed upside-down in order to provide
an early-abort for local changes which allows us to remove a level of
nesting when handling remote changes.