Tests: Allow running async/await in test files
We have previously not been able to use `async/await` in our test files.
This can lead to particularly deep nesting or callback-passing when we
are trying to sequence events for the purpose of the test.

In this change we're adding the `babel-polyfill` to allow us to use
those constructs and _linearize_ the async flows.