Fix: Disconnect before reconnecting to prevent duplication (#90)
In Automattic/simplenote-electron we noticed that after signing out of
the application and signing in again we were getting infinite
duplication of changes. This occurred until we quit the app and
restarted it.

We discovered that when signing back in the Simperium client library was
opening a new WebSocket connection without closing the old one. Whenever
the server sent back changes the client got them twice and tried to
update the content twice, which then cascaded duplication without end.

In this patch we're closing the existing connection if it's already open
when we call `connect()`. This probably fixes defects we haven't
associated with this change.