Play: Inline single-use functions in `channel.js`
There are several helper functions which get called in `channel.js` by
only a single function. In this patch we've inlined those functions in
order to make the sequence of logic and fetching more obvious.

This isn't designed specifically to change the codebase as much as build
a better understanding of the asynchronous points in the channel where
concurrency issues could arise.

This patch may or may not be considered for merging.