Fix: Stop reapplying changes which have already been applied
Co-authored by @beaucollins

This fix finds itself in the midst of a struggle to find the root cause
of the so-called "ghost-writing" or "infinite-duplication" bug. Under
certain relatively-rare and hard-to-reproduce circumstances we find a
given change reapplied as quickly as the library can communicate with
the server. This tends to destroy entities when transforming the change
against itself.

In this patch we're applying two techniques to prevent cycles in our
change update flow: stop applying any patch whose end version is below
our ghost's version; and make sure we ackwnoledge a queued change if
the server sends it back as a duplicate change.. If we have a newer
ghost than the change then it implies that the change has already been
applied. Likewise, if we submit a patch the server has already received
then we should flush any local buffers holding on to it.

These two changes will prevent a subset of possible ways that we can
trigger ghost-writing but it's unclear how significant that subset is
relative to all cases. We know that in some cases the local client
library continues to send changes to the server without updating its
local `sv` and it does not appear like this change will address those
cases.